organise components folder
