+++ src/lib/components/creative/MountainCanvas.svelte
