+++ src/lib/components/creative/OpenCloseBook.svelte
