+++ src/lib/components/layout/CarouselGallery.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { fade } from "svelte/transition";
     import { onMount } from "svelte";
-    import SeekButtons from "./SeekButtons.svelte";
+    import SeekButtons from "../utils/SeekButtons.svelte";
 
     export let darkBackground = false;
     export let title: string;
