+++ src/lib/components/painting/PaintPalette.svelte
