+++ src/lib/components/painting/PaintingSteps.svelte
