+++ src/lib/components/posts/Footnote.svelte
