+++ src/lib/components/posts/PostPreview.svelte
