+++ src/lib/components/strava/Strava2024Goal.svelte.ignore
