+++ src/lib/components/strava/StravaBadge.svelte
