+++ src/lib/components/utils/FaIcon.svelte
