+++ src/lib/components/utils/ZoomImage.svelte
