+++ src/posts/crafting-grand-archway_pt1.md
@@ -7,10 +7,10 @@ published: true
 icon: institution
 ---
 <script>
-    import ImgHoverToggle from "../lib/components/ImgHoverToggle.svelte";
-    import Footnote from "../lib/components/Footnote.svelte";
-    import FaIcon from "../lib/components/FaIcon.svelte";
-    import CarouselGallery from "../lib/components/CarouselGallery.svelte";
+    import ImgHoverToggle from "../lib/components/utils/ImgHoverToggle.svelte";
+    import FaIcon from "../lib/components/utils/FaIcon.svelte";
+    import Footnote from "../lib/components/posts/Footnote.svelte";
+    import CarouselGallery from "../lib/components/layout/CarouselGallery.svelte";
     import ImgGrid from "../lib/components/layout/ImgGrid.svelte";
 </script>
