+++ src/posts/crafting-grand-archway_pt2.md
@@ -8,9 +8,9 @@ pinned: true
 icon: institution
 ---
 <script>
-    import ImgHoverToggle from "../lib/components/ImgHoverToggle.svelte";
-    import Footnote from "../lib/components/Footnote.svelte";
-    import FaIcon from "../lib/components/FaIcon.svelte";
-    import CarouselGallery from "../lib/components/CarouselGallery.svelte";
+    import ImgHoverToggle from "../lib/components/utils/ImgHoverToggle.svelte";
+    import Footnote from "../lib/components/layout/Footnote.svelte";
+    import FaIcon from "../lib/components/utils/FaIcon.svelte";
+    import CarouselGallery from "../lib/components/layout/CarouselGallery.svelte";
     import ImgGrid from "../lib/components/layout/ImgGrid.svelte";
 </script>
