+++ src/routes/projects/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import type { PageData } from "./$types";
-    import ProjectPreview from "../../lib/components/ProjectPreview.svelte";
+    import ProjectPreview from "../../lib/components/posts/ProjectPreview.svelte";
     export let data: PageData;
 
     const getNumberOfPosts = (projectId: string): number => {
